Create README.md
Hotel cancellations have been rising in the past few years due to the presence of online websites (OTA) like Booking.com, expedia.com, etc. Numerical data shows that with the exception of 2018, every single channel has observed a marked increase in cancellation rate YoY. And, even in 2018, the number was 7.1 points above 2014.
OTA´s are encouraging customers to cancel by providing services like book now and cancel later, free of charge, whenever you want. This results in customers booking more than one hotel and making a final decision for their stay later on, hence resulting in cancellations. Examples of the impact of cancellations on a hotel:
• Loss of revenue when they cannot resell the room.
• Additional costs of distribution channels by increasing commissions or paying for publicity to help sell these rooms.
• Lowering prices last minute, so they can resell a room, resulting in reducing profit margin.
So, what can hotels do to reduce this uncertainty and maximize their product and
revenue? A lot can be done with revenue management techniques when it comes to
rates restrictions, like increasing the number of days until the customer can cancel
their booking free of cost, hence giving you more time to resell the room. But now due
to the presence of competition, we need to monitor these policies so that we don’t
loss customers due to stricter cancellation policies compared to our competitors
Therefore, it would seem that we have a complex problem and not a viable solution.
However, now we can use data science and machine learning techniques, to accurately predict which individual and specific reservations are going to cancelled.